requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
PyJWT==2.9.0
pytz==2024.2
spyne==2.14.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.2
